// plio-api/plio/apis/OrderController.ts
/* eslint-disable */

/**
 * Plio.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { createEndpoint, normalizePath } from '@1023-ventures/darri-core';

import {
    Order,
    OrderFromJSON,
    OrderToJSON,
} from '../models/Order';
import {
    ProblemDetails,
    ProblemDetailsFromJSON,
    ProblemDetailsToJSON,
} from '../models/ProblemDetails';

// Data Transfer Objects for individual endpoints

type DeleteUrlParams = {
    /** id  true*/
    id: string;
}

type DeleteQueryParams = {
    /**  */
    concurrency?: string;
}

type CreateBody =
    /**  */
    & Order;

type CreateUrlParams = {
    /** reload  true*/
    reload: boolean;
}

type ReadUrlParams = {
    /** id  true*/
    id: string;
}

type SparseUpdateBody =
    /**  */
    & string;

type SparseUpdateUrlParams = {
    /** id  true*/
    id: string;
    /** reload  true*/
    reload: boolean;
}

type UpdateBody =
    /**  */
    & Order;

type UpdateUrlParams = {
    /** reload  true*/
    reload: boolean;
}

/**
 * no description
 */
export const OrderController = {

    /**
     */
    _delete: createEndpoint<boolean, never, DeleteUrlParams, DeleteQueryParams>({
        method: 'DELETE',
        url: normalizePath('/plio/api/Order/{id}')
    }),
    /**
     */
    create: createEndpoint<Order, CreateBody, CreateUrlParams, never>({
        method: 'POST',
        url: normalizePath('/plio/api/Order')
    }),
    /**
     */
    read: createEndpoint<Order, never, ReadUrlParams, never>({
        method: 'GET',
        url: normalizePath('/plio/api/Order/{id}')
    }),
    /**
     */
    readAll: createEndpoint<Array<Order>, never, never, never>({
        method: 'GET',
        url: normalizePath('/plio/api/Order')
    }),
    /**
     */
    sparseUpdate: createEndpoint<Order, SparseUpdateBody, SparseUpdateUrlParams, never>({
        method: 'PATCH',
        url: normalizePath('/plio/api/Order/{id}')
    }),
    /**
     */
    update: createEndpoint<Order, UpdateBody, UpdateUrlParams, never>({
        method: 'PUT',
        url: normalizePath('/plio/api/Order')
    }),
}